requirements.txt
Django
gunicorn
django-heroku
django-allauth
Pillow
django-crispy-forms
sklearn
numpy
pandas
djangorestframework